fix(proxy): wire translator/StreamGuard/model_locks into routing loop + clean clippy
routing-failover-core-fix 任务收尾：

接线（R1-R4）：
- 新增 proxy/translate.rs：请求 build_translated_request_body (from→to)、
  响应 translate_response_body (to→from)，跨协议转换真正生效
- 新增 proxy/sse.rs：SseLineDecoder + translate_stream，流式响应
  走 StreamGuard 首块探测后逐 chunk 转发，stream_started 状态正确
- model_locks：selector 回调实查 DB，404/model_not_found 写模型锁
- 新增 persist_cooldown helper：冷却时长写 endpoints 表 (cooldown_until/
  last_failure_at/last_error_kind)，重启不丢

死代码清理（R5）：
- final_result (恒 None) → last_mapping: Option<ModelMappingResult>
- body_hash_sync (let _ = no-op) → 删除，改写后重算 hash
- is_stream = method==POST → helpers::is_streaming(&req_json)

附带修复真实逻辑 bug：
- error.rs should_failover: 501 被 500..=599 区间吞掉误判为可退避，
  显式非退避分支上移到 5xx 区间之前

clippy -D warnings 全量清理（27→0）：
- API 预留 dead code（DAO/capability/failover/error/auth_injector）：
  模块级 #![allow(dead_code)] + ///→//! 模块文档
- 真实机械 lint：manual_strip→strip_prefix、type_complexity→ByteStream
  别名、too_many_arguments allow、doc_lazy_continuation 修缩进

质量门禁：cargo fmt/check/clippy -D warnings 全绿，79 测试通过，tsc --noEmit EXIT=0

+++ src-tauri/src/http/proxy/failover.rs
@@ -1,7 +1,10 @@
-/// 故障转移引擎。
-///
-/// 管理端点切换状态、重试计数、冷却逻辑和回退链。
-/// `FailoverState` 提供状态追踪和辅助方法，`proxy_request` 内部驱动路由主循环。
+//! 故障转移引擎。
+//!
+//! 管理端点切换状态、重试计数、冷却逻辑和回退链。
+//! `FailoverState` 提供状态追踪和辅助方法，`proxy_request` 内部驱动路由主循环。
+//!
+//! `RouteAttempt` 与部分重试计数辅助方法为路由管理层预留，尚未在主循环全量接线。
+#![allow(dead_code)]
 use std::collections::{HashMap, HashSet};
 
 use serde::Serialize;
@@ -179,7 +182,7 @@ impl FailoverState {
             ProxyErrorKind::UpstreamError(429) => {
                 // 尝试从 error.message 解析 Retry-After
                 let retry_after = parse_retry_after(&error.message).unwrap_or(5);
-                retry_after + constants::COOLDOWN_429_RETRY_AFTER_BUFFER_SECS as i64
+                retry_after + constants::COOLDOWN_429_RETRY_AFTER_BUFFER_SECS
             }
             ProxyErrorKind::UpstreamError(503) | ProxyErrorKind::UpstreamError(529) => {
                 // 指数退避：2^switch_count 秒，最大 300s
